requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
watchfiles==0.21.0
jaclang==0.8.4
jac-splice-orc  
python-dotenv==1.0.0
pygame
"pydantic>=2.7,<2.9" "pydantic-core<2.23"
mtllm
-e .
transformers
datasets
accelerate
torch
torchvision
peft
jupyter
Pillow


